+++ src/Pim/Bundle/CatalogBundle/Entity/ProductAssociation.php
@@ -2,7 +2,6 @@
 
 namespace Pim\Bundle\CatalogBundle\Entity;
 
-use Doctrine\ORM\Mapping as ORM;
 use Doctrine\Common\Collections\ArrayCollection;
 use JMS\Serializer\Annotation\ExclusionPolicy;
 use Pim\Bundle\CatalogBundle\Entity\Association;
@@ -16,67 +15,32 @@ use Pim\Bundle\CatalogBundle\Model\ProductInterface;
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  *
- * @ORM\Table(name="pim_catalog_product_association")
- * @ORM\Entity(repositoryClass="Pim\Bundle\CatalogBundle\Entity\Repository\ProductAssociationRepository")
- *
  * @ExclusionPolicy("all")
  */
 class ProductAssociation
 {
     /**
      * @var integer
-     *
-     * @ORM\Column(name="id", type="integer")
-     * @ORM\Id
-     * @ORM\GeneratedValue(strategy="AUTO")
      */
     protected $id;
 
     /**
      * @var Association $association
-     *
-     * @ORM\ManyToOne(targetEntity="Pim\Bundle\CatalogBundle\Entity\Association")
-     * @ORM\JoinColumn(name="association_id", nullable=false, onDelete="CASCADE", referencedColumnName="id")
      */
     protected $association;
 
     /**
      * @var ProductInterface $owner
-     *
-     * @ORM\ManyToOne(targetEntity="Pim\Bundle\CatalogBundle\Model\ProductInterface", inversedBy="productAssociations")
-     * @ORM\JoinColumn(name="owner_id", nullable=false, onDelete="CASCADE", referencedColumnName="id")
      */
     protected $owner;
 
     /**
      * @var ProductInterface[] $products
-     *
-     * @ORM\ManyToMany(targetEntity="Pim\Bundle\CatalogBundle\Model\ProductInterface")
-     * @ORM\JoinTable(
-     *     name="pim_catalog_product_association_product",
-     *     joinColumns={
-     *         @ORM\JoinColumn(name="productassociation_id", referencedColumnName="id", onDelete="CASCADE")
-     *     },
-     *     inverseJoinColumns={
-     *         @ORM\JoinColumn(name="product_id", referencedColumnName="id", onDelete="CASCADE")
-     *     }
-     * )
      */
     protected $products;
 
     /**
      * @var Group[] $groups
-     *
-     * @ORM\ManyToMany(targetEntity="Pim\Bundle\CatalogBundle\Entity\Group")
-     * @ORM\JoinTable(
-     *     name="pim_catalog_product_association_group",
-     *     joinColumns={
-     *         @ORM\JoinColumn(name="productassociation_id", referencedColumnName="id", onDelete="CASCADE")
-     *     },
-     *     inverseJoinColumns={
-     *         @ORM\JoinColumn(name="group_id", referencedColumnName="id", onDelete="CASCADE")
-     *     }
-     * )
      */
     protected $groups;